# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Interactive API documentation for the Local Discovery App.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Empre Backend API
  version: "1.0"
paths:
...
  /api/images/{id}:
    get:
      description: Fetch image binary data using its secure UUID mapping
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/png
      - image/jpeg
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get image by secure ID
      tags:
      - Media
  /api/images/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image to S3 and get a secure UUID mapping
      parameters:
      - description: Image File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a general image
      tags:
      - Media
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"